First cut at the hub page. I needed to create a gdp_custom.css that gets included after bootstrap since custom.css is included before because it is part of the usgs template.
Added methods to JobModel to provide the process model which has been selected and also a method to return the process inputs that should be shown on the process page.

I now remove and reinitialize the AlgorithmConfig View when changing the Algorithm name. this is cleaner as I can initialize the process variables when switching to a new algorithm without worrying about signal handlers firing in AlgorithmConfigView.